build_conf: introduce a new configuration file processor
Present configuration processor in build_conf_processor have a number
of pitfalls, including:

 - It can't report where in YAML file it stumbled on an error
 - There is a bug with when reporting unknown variables - it produces
   very generic error which can't show what problem is and with which
   variable
 - It presents configuration as untyped dicts and lists

This patch adds a new processor that is free of mentioned issues. It
employs pyyaml API for Nodes and Marks which allows us to typecheck
data in YAML file and always know location for every piece of input
data. Also it wraps all this into objects.

This greatly enhances user experience, as now user at least can get
location of error in YAML file.

To be able to work with other components of moulin,
MoulinConfiguration class provides get_document() method, which
produces good old dict of lists and dicts, which then can be used by
builders and fetchers.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>