GitFetcher: add revision to the checkout_stamp value.
This is necessary to fetch the same layer with different revisions,
now, every component may have dedicated revision of the layer.
But, the  error when one revision is replaced with another one is possible.
It happens if one layer is mentioned twice in the source list of the component
with the different revision.
This fix raise the exception with the text :
"Layer [layer url] already exists but with another revision, expected [last revision].
Maybe layer [layer url] is mentioned twice in the soure list."

Signed-off-by: Ihor Usyk <ihor_usyk@epam.com>