build_generator: switch to extended YAML parsing API
This is quite big patch, as it is not feasible to convert each
component separately. Basically we are switching from old "dict of
dicts of list" approach for storing build configuration to use of
pyyaml.Node interface. This allows us to report errors with location
in yaml file, which is more user-friendly.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>