moulin: switch to YamlValue API
This huge patch switches the whole code base to a new API based on
YamlValue. Unfortunately we can't split it into multiple smaller
patches as this will break build.

Good news is that this patch does not produce any functional changes,
this is just 1:1 mapping from old API not a new one.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>