moulin: switch to "deps" dependency tracking
Ninja supports two ways to discover dynamic dependencies:

 - dyndep mechanism - it is used to discover mandatory dependencies
   *before* build
 - deps mechanism - it is used to discover optional dependencies
   *during* build. It is used mostly to track dependencies between
   C source files and header files they include

Prior to this patch, Moulin used the dyndep mechanism. But it appears
that it does not cope will with removing or renaming files. As those
type of dependencies are mandatory, Ninja will complain that it does
not know how to build removed file.

So, only one way to fix this is to move to "deps" mechanism. This is
exactly what this patch does.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>