fetchers: git: Use ssh in batch mode
If $git_url uses ssh protocol and is not in .ssh/known_hosts,
the ssh client starts to interact with a user.
But this interaction is not visible and looks
just like the ninja hung on some operation.
So we have to use ssh in BatchMode to throw
an error instead of asking invisible questions.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>