Revert "rouge: Add optional compress flag"
This reverts commit afefdeb1cc585e5b5fcc6e9c445d032b6f7da040.

After more detailed testing and rethinking it turned out that rouge is
not the right tool for this task. This implementation does not fit well
with moulin's ninja rules, and does not fit well with rouge's philosophy
of being a tool for creating an image. So the better thing to do would
be to revert this, and rewrite the functionality in ninja rules.

Signed-off-by: Mykyta Poturai <mykyta_poturai@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>