Revert "rouge: Add optional bmap creation"
This reverts commit a9e450dfc0b7e2a1dfed40c70faad8cadb5ca1ce.

After more detailed testing and rethinking it turned out that rouge is
not the right tool for this task. This implementation does not fit well
with moulin's ninja rules, and does not fit well with rouge's philosophy
of being a tool for creating an image. So the better thing to do would
be to revert this, and rewrite the functionality in ninja rules.

Signed-off-by: Mykyta Poturai <mykyta_poturai@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>