rouge: fix partition alignment
Do not add 1 additional sector to the end of each partition, as this
would unnecessarily add space between partitions that are already
aligned to DEFAULT_ALIGNMENT. Selection of 'next sector' is already
done by start + size calculation.

This avoids wasting space and allows to align the first partition
exactly to DEFAULT_ALIGNMENT (and not to DEFAULT_ALIGNMENT + 2048,
as it is now).

Signed-off-by: Francesco Valla <francesco.valla@mta.it>
Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>