rouge: add possibility to specify size of image
For some cases we need to have image of certain size
despite sizes of internal partitions.

This PR adds `image_size` field to image descriptor.
If it is not specified or zero, then image is created taking
into account internal partitions.
If actual size of partitions is bigger than `image_size` -
error is raised, that we can't create image of specified size.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>