builders: archive: introduce tar's base dir
This commit solves an issue that archives created by the archive builder
contain paths to files relative to build-dir. As result, paths may
contain a long list of build-related directories, and in some cases, this
is not desirable.
A possible solution is to specify desired starting point as build dir,
but this will put the resulting archive into the same directory.
So, the best way to get an archive in the top directory and get rid of
intermediate directories - is to specify the base directory for `tar`
through the `-C` option.

That's why this commit introduces the `base_dir` field for the archive
builder. The `base_dir` is an optional field, and it has '.' value if
not provided.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>