rouge: add GPT image wrapper
gpti.py shall be used as a replacement for sfdisk.py. It ulilizes
gpt-image primitives like Table, Geometry and Partition to create a
custom disk image.

Yet we want to avoid use of Disk interface from gpt-image. One [minor] reason is
that it does not overwrite existing image file, second [major] is that
it zeroes the full image, but we'd like to keep it sparsed.

Signed-off-by: Sergiy Kibrik <sergiy_kibrik@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>