raise exception when have ambiguousness of revisions
yaml file may have some ambiguousness of revisions, for example
.....
domd:
    sources:
      - type: git
        url: "https://github.com/good/meta-good"
.....
--ENABLE_SOMETHING
      default: no
      ......
         components:
            domd:
              sources:
                - type: git
                   url: "https://github.com/good/meta-good"
                   rev: main
Here we have one repository but with the two different revisions.
At this moment, moulin checks if url has been fetched and skip all the
overrides. It means that build may have unexpected result.
The patch check if the repos exists and if the revisions are the same.
If revisions are not the same - remove build file, throws the exception
and stop the running.

Signed-off-by: Ihor Usyk <ihor_usyk@epam.com>