builders: Added escape of 'env' parameter
There is an issue with passing the "EXT_MODULES" parameter from the
configuration moulin yaml file to the build.sh context.

The 'android_kernel' builder didn't escape double quotes, while
wrapping the 'env' parameter in additional double quotes like this:

bash -c "...env..."

As a result, an incorrect syntax was generated in the build.ninja file.

To generate a functional build.ninja file, it was necessary to take
this aspect into account in the moulin.yaml file and add additional
escaping there.

This approach is ambiguous since it requires the user to know the
implementation details of each moulin builder.

To address this issue, a simple escape function was added to all
builders that support the 'env' parameter. It escapes double quotes to
ensure proper handling of the 'env' variable.

The issue has been resolved for the following builders:
- android_kernel builder
- android builder
- zephyr builder

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>