fetchers: disable prompt usage by git
This commit addresses issue with git interacting with
the user through the console. For example, HTTPS access
to the private repo.
We use git inside the hidden console, so user sees only
hanging fetch.

So, we instruct the git to not interact with the user by the command
`GIT_TERMINAL_PROMPT=0`. As result, git fails cloning and inform
user that asking name/password is not possible.
This should push the user to use script-friendly authorization,
like ssh.

Suggested-by: Artem Mygaiev <artem_mygaiev@epam.com>
Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>