git.py: Fail on different branches for same repo
  doma:
    sources:
      - type: git
        url: https://github.com/xen-troops/meta-example.git
        rev: "ABCDEF"
  ...
  domu:
    sources:
      - type: git
        url: https://github.com/xen-troops/meta-example.git
        rev: "branch_A"

For doma and domu, the same GitHub repository is specified, but with
different branches.

In the case of such an erroneous YAML file, Moulin will not highlight
this error. Consequently, the user will also spend significant effort
to understand the reason for the incorrect functioning.

This patch adds an additional check to the git.py file. Now, if the
repository branches don't match, an error is displayed in the console,
and the generation of the build.ninja file is halted.

Co-Authored-By: Ihor Usyk <ihor_usyk@epam.com>
Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>