yocto.py: Support aliases for specifying 'layers'
...
      common_layers: &COMMON_LAYERS
            - "layer1"
            - "layer2"
...
components:
    comp0:
         builder:
             layers:
                 - *COMMON_LAYERS
                 - "layer3"
                 - "layer4"

The issue is that if an alias appears inside the 'layers' key, the
alias reaches us as an empty node without a value but with a child.
It's the same as writing:

layers:
     -
        - "nested_layer1"
        - "nested_layer2"
     - "layer3"
     - "layer4"

And there is no way to step over this on the yaml level, e.g. using
some other syntax.

To make the "layers" component combinable with a variable, as in
the example above, the changes were made to the yocto.py file.
The function _flatten_layers has been added. It processes an input YAML
structure, extracting the names of individual layers as strings and
appending them to the resulting list.
The function will return:
['nested_layer1', 'nested_layer2', 'layer3', 'layer4']

Now the above empty node is unfolded for this specific case, and the
proper data is fetched by the moulin.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>