git.py: added directory tracking to the Moulin git builder
Change the behavior of the git fetcher so that it doesn't consider two
or more occurrences of the same git repository in the moulin *.yaml
configuration file as an error if the duplicated repository is cloned
into different sub-folders within the build directory.
For example:

  domd:
     sources:
      - type: git
        url: "https://github.com/xen-troops/meta-example.git"
        rev: ABCDEF
        dir: "dir1"
  domu:
    sources:
      - type: git
        url: "https://github.com/xen-troops/meta-example.git"
        rev: branch_A
        dir: "dir2"
Above, identical repositories have different branches. But, also, they
have different 'dir' parameters. So that would not be considered as an
error.

But if they would have identical directories - the error still will be
triggered.

This fix adds the value, that is specified by the user as a 'dir'
parameter of the moulin's 'git' fetcher to the build.ninja file
markers, created during the execution of 'git clone' and 'git checkout'
operations. After this change, if repositories with the same name have
different directories, unique clone_stamp and checkout_stamp file
markers will be created for each repository. In this case, revision
checking of such repositories will not be performed, which is the
desired behavior.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Fixes: 19d2ec09c436 ("git.py: Fail on different branches for same repo")