yocto.py: update layers in bblayers.conf
If after building a moulin component (e.g., dom0), layers in the
Yaml file are modified, these changes will not be reflected in
bblayers.conf.

This fix addresses this issue.
The following dependencies were modified:

Dependencies of the dom0 target before changes

+------------------------------------+
| ./yocto/build-dom0/conf/local.conf |
+------------------------------------+
           | (yocto_add_layers)
           |
           v
+---------------------------+
| ./stamps/yocto-build-dom0-|
| yocto-layers              |
+---------------------------+
           | (yocto_update_conf)
           |
           v
+-------------------------------------+
| ./yocto/build-dom0/conf/moulin.conf |
+-------------------------------------+
           |
           | (yocto_build)
           v
       +-------------+
       | yocto_build |
       +-------------+

Dependencies of the dom0 target after changes

+------------------------------------+
| ./yocto/build-dom0/conf/local.conf |------+
+------------------------------------+      |
           |                                |
           |                                |
           |                                |
           | (yocto_update_conf)            | (yocto_add_layers)
           |                                |
           |                                V
           |                  +---------------------------+
           |                  | ./stamps/yocto-build-dom0 |
           |                  | yocto-layers              |
           |                  +---------------------------+
           |                                |
           V                                |
+-------------------------------------+     |
| ./yocto/build-dom0/conf/moulin.conf |     |
+-------------------------------------+     |
           |                                v
           |                         +-------------+
           +------------------------>| yocto_build |
                                     +-------------+

Also, the fix touches the yocto_add_layers ninja rule. Now, it checks
the set of the provided layers against the cached data. If a cache file
exists, the logic compares the provided layers against the cached ones.
If there are differences between two data sets, the rule removes all
previous layers and adds the new set of layers. If the file doesn't
exist yet, the logic adds the layers to the Yocto and updates the
cache of the added layers for further usage. After application of
this patch, if you remove or add layers in the moulin *.yaml
configuration file, it will be propagated to the bblayers.conf file.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>