unpack.py: add dynamic dependecies to the unpack fetcher
Enable building the build.ninja even when the moulin YAML configuration
specifies an archive that is not present in the file system.

Detection of the missing archive error will now occur during the ninja
invocation, specifically when building a ninja target that unpacks the
archive.

Implement functionality to generate dyndep files for archives. This
will ensure that Ninja will unpack archive if it was updated or
replace any missing files. This functionality is achieved by calling
moulin.fetchers.archive module using python. In this way we can use
the services provided by Moulin to generate Ninja dyndep file.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>
[Volodymyr: added dyndep generator via "python -m" call]
Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>