Revert "yocto.py: update layers in bblayers.conf"
This reverts commit dc4ef76a1195fb2f95649efa3075737daf032c24.

This commit introduced a bug, that prevents the build system adding
back layers when "conf" directory is removed. User visible impact is
following:

ERROR:  OE-core's config sanity checker detected a potential misconfiguration.
    Either fix the cause of this error or at your own risk disable the checker (see sanity.conf).
    Following is the list of potential problems / advisories:

    MACHINE=NNNNN is invalid. Please set a valid MACHINE in your local.conf, environment or other configuration file.

Root cause for this bug is that Ninja now makes assumption about
layers.conf based on stamp file, not on modification
time of layers.conf itself.

This bug is much more annoying than issue that this commit tried to
resolve, so it is better to revert the commit until proper fix is
found.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>