rouge: workaround changes in gpt-image
gpt-image is package for working with GPTs. We use it to create
partition layout in rouge. To properly write sub-images into final
image, we need to know offset for every partition. To do this, we
read "first_lba" field in Partition object. Problem is that in
gpt-image 0.9.0 this property contains zero and we need to read
"first_lba_staged" property instead.

Also, we need to support old installation, so we just can't replace
"first_lba" with "first_lba_staged" in the code. Instead we should try
the new field first and then fall back to the old one.

Tested with both gpt-image 0.9.0 and gpt-image 0.8.1

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Reviewed-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Tested-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>