rouge: add Hybrid MBR support for GPT
Hybrid MBR is kind of accepted hack, that defines (subset) of
partitions both in MBR and in GPT. It is primarily used in cases when
we want to use GPT, but some part of the system (like bootloader)
supports only MBR. This thing is not standardized so different OSes
can react differently to this it. Thus, it should be used only if
there are no other options.

This patch was made with intention to be able to create Raspberry Pi
Zero boot images.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Tested-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>