rouge: Add support for resizing raw images
Add new "resize" option to the raw_image block, which allows rouge
to resize the raw_image to the size set in the block. It is enabled
by default and can be disabled manually.

This is useful when the image size needs to be changed frequently, or
multiple images need to be created from the same source files, as it
allows to skip the rootfs rebuild step.

Images are copied to a temporary directory before resizing to preserve
the original build artifacts. The temporary directory is created in the
current working directory to prevent filling the system's /tmp with
large images.

Signed-off-by: Mykyta Poturai <mykyta_poturai@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Reviewed-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>