logs: Add colored output
Make the log output colored depending on the log level. This will make
it much easier to quickly spot errors and warnings in the output, and
also looks cooler.

Control characters for colors are only added if the output is a TTY, so
the output can be redirected to a file without any issues.

Signed-off-by: Mykyta Poturai <mykyta_poturai@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>