rouge: add `sparse` option to images
This option is enabled by default for compatibility with previous
version. Also, this is sane value. But there are times, when `dd`
sparsed region detector behaves too greedy and emits un-allocated
regions even when in the original file those regions were populated
with zeroes. This is not a problem in day-to-day life, but can cause
issues when writing to un-initialized flash storages. To mitigate this
issue, `sparse` option is added to block description. This option
forces `dd` to generate un-sparsed regions, which may help in some
use-cases.

Signed-off-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>