Provide a clear message if the default option is not specified
This change handles the case when "default: true" is not provided
for some option in the YAML file.

Now, in this scenario, a clear "YAMLProcessingError" is raised,
containing detailed and understandable information about the
misconfiguration.

Added pytests for the following scenarios:
-- YAML parameter does not contain a "default: true" variant
(negative case)
-- YAML parameter contains a "default: true" variant (positive case)

Also, the unit test "test_config.py" was updated because
"default: true" is required for all parameters in the YAML
configuration.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>