yocto: change the additional_deps handling
The option `additional_deps` is part of the yocto builder,
and thus applies to the whole build target, if used.

At the same time, yocto builder has to do a series of steps
described as a few build targets internally.

This commit changes the creation of the build targets so
that the external dependency must be satisfied before the
very first build target.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>